`ProductRequestData`: changed tests from manual dictionary comparisons to snapshot testing
Future refactor is removing manual JSON encoding, and this is easier to maintain without needing to hardcode key names.
